+++ packages/grimmcraft-core/srcs/grimmcraft_core/entity/__init__.py
@@ -5,12 +5,11 @@ from __future__ import annotations
 from grimmcraft_core.entity.core_entity import CoreEntity
 from grimmcraft_core.entity.mob import Mob
 from grimmcraft_core.entity.npc import Npc, Trade, VillagerProfession
-from grimmcraft_core.entity.player import GameMode, Player
+from grimmcraft_core.entity.player import Player
 
 __all__ = [
     "CoreEntity",
     "Player",
-    "GameMode",
     "Mob",
     "Npc",
     "Trade",